requirements.txt
# fastapi==0.110.0
# uvicorn[standard]==0.29.0
# pymongo==4.6.1
# python-dotenv==1.0.1
# pillow==10.2.0
# torch==2.2.0
# torchvision==0.17.0
# tqdm==4.66.2
# scikit-learn==1.4.1.post1
# pandas==2.2.1
# numpy==1.26.4
# requests==2.31.0
fastapi
uvicorn
pillow
torch
ultralytics
pymongo
python-dotenv
